Move the toggle mute button to its own component.
