+++ multiplayer/src/components/GamePage.tsx
@@ -1,17 +1,12 @@
 import { faCopy } from "@fortawesome/free-regular-svg-icons";
-import {
-    faCheck,
-    faVolumeHigh,
-    faVolumeMute,
-} from "@fortawesome/free-solid-svg-icons";
+import { faCheck } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { useContext, useEffect, useState } from "react";
-import { Button } from "react-common/components/controls/Button";
-import { setMute } from "../state/actions";
 import { AppStateContext, dispatch } from "../state/AppStateContext";
 import ArcadeSimulator from "./ArcadeSimulator";
 import Presence from "./Presence";
 import Reactions from "./Reactions";
+import ToggleMuteButton from "./ToggleMuteButton";
 
 export interface GamePageProps {}
 
@@ -30,10 +25,6 @@ export default function Render(props: GamePageProps) {
         }
     };
 
-    const toggleMute = () => {
-        dispatch(setMute(!state.muted));
-    };
-
     useEffect(() => {
         if (copySuccessful) {
             let resetCopyTimer = setTimeout(() => {
@@ -44,10 +35,6 @@ export default function Render(props: GamePageProps) {
             };
         }
     }, [copySuccessful]);
-
-    useEffect(() => {
-        pxt.runner.currentDriver()?.mute(state.muted);
-    }, [state.muted]);
 
     return (
         <div>
@@ -60,16 +47,7 @@ export default function Render(props: GamePageProps) {
                 <div className="tw-flex tw-flex-col tw-items-center">
                     {state.playerSlot && <ArcadeSimulator />}
                     <div className="tw-flex tw-flex-row tw-w-full tw-items-center tw-justify-between tw-mt-1">
-                        <Button
-                            leftIcon={
-                                state.muted
-                                    ? "fas fa-volume-mute"
-                                    : "fas fa-volume-up"
-                            }
-                            title={lf("Toggle Mute")}
-                            className="tw-border-2 tw-border-slate-400 tw-border-solid tw-py-2 tw-pl-2 tw-pr-1 tw-bg-slate-100 hover:tw-bg-slate-200 active:tw-bg-slate-300"
-                            onClick={toggleMute}
-                        />
+                        <ToggleMuteButton />
                         <div className="tw-justify-self-center">
                             {state.gameState?.joinCode && (
                                 <div>
